pyproject.toml
[project]
name = "sktr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
classifiers = ["Private :: Do Not Upload"]
dependencies = [
  "jaxtyping>=0.3.2",
  "pydantic>=2.11.7",
  "pymilvus>=2.6.0",
  "timm>=1.0.17",
  "torch>=2.7.0",
  "tqdm>=4.67.1",
]

[project.scripts]
train = "sktr.train:train"
server = "server:serve"



[project.optional-dependencies]
server = [
  "fastapi>=0.120",
  "uvicorn>=0.38",
  "watchdog>=6.0.0",
  "blinker>=1.9.0",
  "pyjwt>=2.10.1",
  "sqlmodel>=0.0.27",
  "pydantic[email]>=2.11.7",
]
cpu = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
]
cu128 = [
  "torch>=2.7.0",
  "torchvision>=0.22.0",
]
train = [
  "torchvision>=0.22.0",
  "tensorboard>=2.20.0",

]


tests = [
    "pytest>=8.4.1",
    "matplotlib>=3.10.3",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
]

[tool.uv]
conflicts = [
  [ { extra = "cpu" }, { extra = "cu128" } ],
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"



[tool.pytest.ini_options]
testpaths = ["src/tests/"]
pythonpath = ["src"]
